feature: get all users

+++ routes/api.php
@@ -35,33 +35,34 @@ Route::post('images/products', [ImageController::class, 'store']);
 
 // Products
 Route::get('products', [ProductController::class, 'index']);
-Route::post('products',[ProductController::class, 'store']);
+Route::middleware('auth:api')->post('products',[ProductController::class, 'store']);
 Route::get('products/{id}',[ProductController::class, 'show']);
-Route::put('products/{id}',[ProductController::class, 'edit']);
-Route::delete('products/{id}',[ProductController::class, 'destroy']);
+Route::middleware('auth:api')->put('products/{id}',[ProductController::class, 'edit']);
+Route::middleware('auth:api')->delete('products/{id}',[ProductController::class, 'destroy']);
 Route::get('products/category/{category}',[ProductController::class, 'category']);
 
 // Orders
 Route::get('orders/', [OrderController::class, 'index']);
-Route::post('orders/', [OrderController::class, 'store']);
+Route::middleware('auth:api')->post('orders/', [OrderController::class, 'store']);
 Route::get('orders/{id}', [OrderController::class, 'show']);
 Route::put('orders/{id}', [OrderController::class, 'edit']);
 Route::delete('orders/{id}', [OrderController::class, 'destroy']);
-Route::put('orders/{id}/status', [OrderController::class, 'updateStatus']);
-Route::put('orders/{userId}/', [OrderController::class, 'userOrders']);
+Route::middleware('auth:api')->put('orders/{id}/status', [OrderController::class, 'updateStatus']);
+Route::middleware('auth:api')->put('orders/{userId}/', [OrderController::class, 'userOrders']);
 
 // Product orders
 Route::get('product-orders/', [ProductOrderController::class, 'index']);
-Route::post('product-orders/', [ProductOrderController::class, 'store']);
+Route::middleware('auth:api')->post('product-orders/', [ProductOrderController::class, 'store']);
 Route::get('product-orders/{id}', [ProductOrderController::class, 'show']);
-Route::put('product-orders/{id}', [ProductOrderController::class, 'edit']);
-Route::delete('product-orders/{id}', [ProductOrderController::class, 'destroy']);
+Route::middleware('auth:api')->put('product-orders/{id}', [ProductOrderController::class, 'edit']);
+Route::middleware('auth:api')->delete('product-orders/{id}', [ProductOrderController::class, 'destroy']);
 
 Route::get('rating/', [RatingController::class, 'index']);
-Route::post('rating/', [RatingController::class, 'store']);
+Route::middleware('auth:api')->post('rating/', [RatingController::class, 'store']);
 Route::get('rating/{id}', [RatingController::class, 'show']);
-Route::put('rating/{id}', [RatingController::class, 'edit']);
-Route::delete('rating/{id}', [RatingController::class, 'destroy']);
+Route::middleware('auth:api')->put('rating/{id}', [RatingController::class, 'edit']);
+Route::middleware('auth:api')->delete('rating/{id}', [RatingController::class, 'destroy']);
 
 //User
-Route::get('user/{id}/orders', [UserController::class, 'userOrders']);
+Route::middleware('auth:api')->get('user/{id}/orders', [UserController::class, 'userOrders']);
+Route::middleware('auth:api')->get('user/list', [UserController::class, 'index']);
